Updated Unit Tests; Add Black and autoflake; Purged some unit tests
Removed some unit tests as they used HTML parsing, which caused a ton of issues with our UI rewrite. While we could redo them, it's going to take forever, and I'm only seeing it causing a bunch of wasted time in the future if drastic changes are ever done again. If issues pop up in the future, we'll redo those unit tests to accomodate for those that are now gone.

Also removed the images in 'web/html/images'.