Fixed issue with Repology out of date checker
Previously we only checked repology against the 'pkgname' of any given package. Package names are normalized on Repology though, so we need to handle that by just looping over every package in the Repology package list and checking if the 'srcname' variable matches that on the MPR.

Also added missing user prefixes to commands in Cron script.