Added notices about registration and password reset not working right now
Hetzner is blocking access to ports 25 and 465 for us since we just made our account there, and that's causing issues when sending out emails, which are required for email and password resets.

We'll move the MPR back to Linode in a few days if we can't get this resolved in time.