Convert to new Git repository layout
We're using a new layout for Git reposities so that we can better manage certain features (like version tags on each commit). This required us to have a different repository for each package instead of just using different tags, so we had to completely redo the layout.

The MPR instance has already been migrated to the new layout, it's just awaiting this change to make it all work properly.