Disable namespaces in SSH Git interface [TEST SKIP]
We use separate repositories for everything now, so we don't need them. They also bring us to separate histories when using the HTTP interface, so we need to stop using them because of that too.

We just spent over two hours getting this fixed :P. Sure glad we found the culprit though.